Изменен алгоритм парсинга: теперь не используется расчет сразу после парсинга каждой следующей лексемы. Сначала строится дерево выражений, а затем вычисляется. Это позволяет вынести логику расчета в другую сущность, а операторы и работу с ними в третью

+++ MathExpressionParser.ConsoleApp/Program.cs
@@ -10,7 +10,7 @@ namespace MathExpressionParser
             string inputExpression = string.Empty;
 
             var calculator = new Calculator();
-            double result = default;
+            decimal result = default;
 
             Console.WriteLine("Введите \"q\" если захотите выйти \nВведите выражение");
             inputExpression = Console.ReadLine();
